21211005 V1.2
1.增加SMU Ini時Receive資料流的動作，收掉不必要的回傳訊息
2.修改Script 不用AutoRange，不同26系列設定似乎不同，用2636時AutoRange會量到不對的值，統一拿掉。
3.Bias設定參數型別改 Double(只是想縮大小)
4.把IP設定存進靜態類別(使用可序列畫 字典)
5.校正資料電流部分改用 List
6.電流校正第一檔 增加到 21筆

+++ GUI/frmMsrtSetMPDA.cs
@@ -69,7 +69,7 @@ namespace GUI
                 this._mpdaParameter.DelayTimeSet = (int)(nudDelayTime.Value * 10);
                 this._mpdaParameter.TimeBase = 1;
                 this._mpdaParameter.BiasEnable = chbBiasEnabled.Checked;
-                this._mpdaParameter.BiasVoltage = nudBiasVoltage.Value;
+                this._mpdaParameter.BiasVoltage = (double) nudBiasVoltage.Value;
                 try
                 {
                     this._mPDA.SetAllParameter(this._mpdaParameter);
